chore: Rename BodyResolver to ValueResolver.
