--- src/Http/Factory/Body/BodyResolverInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace CHStudio\Raven\Http\Factory\Body;
-
-interface BodyResolverInterface
-{
-    public function resolve(mixed $value): mixed;
-}
