--- src/Http/Factory/Body/PassThroughValueResolver.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace CHStudio\Raven\Http\Factory\Body;
-
-class PassThroughValueResolver implements BodyResolverInterface
-{
-    public function resolve(mixed $value): mixed
-    {
-        return $value;
-    }
-}
