+++ src/Http/Factory/Resolver/PassThroughValueResolver.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace CHStudio\Raven\Http\Factory\Resolver;
+
+class PassThroughValueResolver implements ValueResolverInterface
+{
+    public function resolve(mixed $value): mixed
+    {
+        return $value;
+    }
+}
