+++ src/Http/Factory/Resolver/ValueResolverInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace CHStudio\Raven\Http\Factory\Resolver;
+
+interface ValueResolverInterface
+{
+    public function resolve(mixed $value): mixed;
+}
